+++ vue/src/board/.charts/empty/EmptyChart.vue
