+++ vue/src/board/.charts/error/ErrorChart.vue
