+++ vue/src/board/.charts/image/ImageChart.vue
