+++ vue/src/board/.charts/line/components/G2Line.vue
