+++ vue/src/board/.charts/line/components/LineLayout.vue
