+++ vue/src/board/.charts/line/components/LineLegends.vue
