+++ vue/src/board/.charts/line/components/smooth.js
