+++ vue/src/board/.charts/text/TextChart.vue
