+++ vue/src/board/types/charts.d.js
